Conflicts were resolved
