+++ cluster-config.yaml
@@ -5,3 +5,6 @@ nodes:
   - role: control-plane
   - role: worker
   - role: worker
+networking:
+  podSubnet: "10.244.0.0/16"
+  serviceSubnet: "10.96.0.0/16"
