+++ integration_k8s_kind_test/files/userConfigMap.yaml
